requirements.txt
Flask
Pillow >= 8.0.1
qrcode
notify2
